requirements.txt
cffi==1.16.0
cryptography==43.0.0
pycparser==2.22
python-dotenv==1.0.1
